Fix remaining compilation errors
- HazardZoneIndicator.kt: Replace MissDirection.LONG/SHORT with THIN/FAT
- ReadinessBreakdownChart.kt: Fix Double/Float type mismatch
- WindIndicator.kt: Add missing rotate import
- IntentTraceView.kt: Add missing AnalyticsEvent branches
- NavCaddyDomainModule.kt: Implement check() method in PrerequisiteChecker

+++ android/app/src/main/java/caddypro/ui/caddy/components/ReadinessBreakdownChart.kt
@@ -188,7 +188,7 @@ private fun MetricBar(
         ) {
             Box(
                 modifier = Modifier
-                    .fillMaxWidth(fraction = (metric.value / 100f).coerceIn(0f, 1f))
+                    .fillMaxWidth(fraction = (metric.value.toFloat() / 100f).coerceIn(0f, 1f))
                     .height(12.dp)
                     .clip(RoundedCornerShape(6.dp))
                     .background(barColor)
